+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 2.1.1
 =====
 
-* (improvement) Replace `.` with `-` in `TaskMetaData::getKey`.
+* (improvement) Replace `.` with `:` in `TaskMetaData::getKey`.
 * (improvement) Change column of `TaskRun::$taskLog` to be not nullable.
 * (internal) Update `21torr/janus` and use `phpunit` instead of `simple-phpunit`.